Made song editable
Based on pull request #1...
The list of frequencies to be played has been moved from the JS to a textbox in the HTML.
Editing the content of the box will change what gets played.
The format also is more human human readable and consistent with the songfile from #1.
If you copy and paste the smb.song file, it should work.