Added a delay function and fixed time being *1000.
There was a bug before, that caused the window to freeze up (seemingly forever). This was due to the fact that the times in milliseconds were being treated as time in seconds. This has been fixed with a *.0001 function.

Also, a block of code for the "delay" function has been added. Delays between same notes in the included demo song have also been added.