Improvements to html, css, and js (#30)
* Improved formatting

* removed unused and unrelated  CSS classes

* removed more unused CSS

* removed red outline around text box

* improved link styling

* Added a delay function and fixed time being *1000.

There was a bug before, that caused the window to freeze up (seemingly forever). This was due to the fact that the times in milliseconds were being treated as time in seconds. This has been fixed with a *.0001 function.

Also, a block of code for the "delay" function has been added. Delays between same notes in the included demo song have also been added.

* fixed description meta

* created var for window log