Implemented Web Workers
Functionality is now taken care of by Web Workers. The advantages of this is that the page will no longer freeze/lock up when the script is working.
This also means that the page status can now update in realtime. (Before, it only updated after the script finished running.) Because it possible for buttons to be clicked during operation, it is now possible to stop script mid-execution.
Song recursion should also be possible now but, has not been implemented.